split graph/table into a different file
