--- src/components/science/sensor.jsx
@@ -1,5 +1,0 @@
-import 'react-resizable/css/styles.css';
-
-export default function sensor() {
-
-}
